requirements.txt
django-model-utils==4.3.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-environ==0.11.2
environ==1.0
coreapi==2.3.3
psycopg2-binary==2.9.9
Django==5.0.1
pytest-cov==4.1.0
django-db-connection-pool==1.2.4
django-rest-swagger==2.2.0
ipdb==0.13.13
pytest==7.4.4
pytest-sugar==0.9.7
pytest-cov==4.1.0
coverage==7.4.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
pytest-django==4.7.0
django_coverage_plugin==3.1.0
